Step Three: Map Scopes to Authorities
As Opaque Token support is the latest addition in Spring Security's
resource server support, OpaqueTokenIntrospector currently fills
multiple roles. Whereas with JWT we'd map authorities with
JwtAuthenticationConverter and validate the token with JwtDecoder,
OpaqueTokenIntrospector does both, possibly to be broken up at a later
date.

Because of that, you still need to delegate to the default opaque token
introspector to validate the token in order to customize the
authorities.